Move the ETL courses endpoint to `/api/internal/courses`
