+++ courses/views/v3/__init__.py
@@ -7,9 +7,8 @@ import re
 
 import django_filters
 from django.conf import settings
-from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ImproperlyConfigured
-from django.db.models import Count, Prefetch, Q
+from django.db.models import Prefetch, Q
 from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
 from drf_spectacular.types import OpenApiTypes
@@ -22,11 +21,7 @@ from drf_spectacular.utils import (
 )
 from rest_framework import mixins, serializers, status, viewsets
 from rest_framework.decorators import api_view, permission_classes
-from rest_framework.permissions import (
-    AllowAny,
-    BasePermission,
-    IsAuthenticated,
-)
+from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.response import Response
 
 from b2b.models import ContractPage
@@ -34,7 +29,6 @@ from courses.api import create_program_enrollments, deactivate_run_enrollment
 from courses.constants import COURSE_KEY_PATTERN, ENROLL_CHANGE_STATUS_UNENROLLED
 from courses.models import (
     Course,
-    CourseRun,
     CourseRunEnrollment,
     Program,
     ProgramEnrollment,
@@ -43,36 +37,18 @@ from courses.serializers.v3.courses import (
     CourseOutlineResponseSerializer,
     CourseRunEnrollmentSerializer,
     CourseVariantRunsResponseSerializer,
-    IngestibleCourseWithCourseRunsSerializer,
 )
 from courses.serializers.v3.programs import (
     ProgramEnrollmentCreateSerializer,
     ProgramEnrollmentSerializer,
 )
 from courses.utils import get_enrollable_courseruns_qs
-from courses.views.utils import Pagination
 from ecommerce.models import Product
 from main import features
 from openedx.api import get_edx_course_outline
-from openedx.constants import EDX_ENROLLMENT_VERIFIED_MODE
 from openedx.exceptions import EdxApiCourseOutlineError
 
 log = logging.getLogger(__name__)
-
-
-class IsEtlUser(BasePermission):
-    """Allow only is_etl flagged users through."""
-
-    message = "Invalid user."
-
-    def has_permission(self, request, view):  # noqa: ARG002
-        """Check the user's is_etl flag."""
-
-        return (
-            request.user
-            and not request.user.is_anonymous
-            and (request.user.is_etl or request.user.is_superuser)
-        )
 
 
 class UserEnrollmentFilterSet(django_filters.FilterSet):
@@ -501,67 +477,3 @@ def get_course_variant_runs(request):
     )
 
     return Response(CourseVariantRunsResponseSerializer(output, many=True).data)
-
-
-class IngestibleCourseViewSet(viewsets.ReadOnlyModelViewSet):
-    """
-    Largely a copy of the v2 CourseViewSet, but with changes for ETL processes.
-
-    The shape of the data returned should match the v2 CourseViewSet for the most
-    part, but we won't support lookups or filtering or anything and the prefetch
-    setup will be reduced (since this won't have to care about what the requesting
-    user's contracts are for instance).
-    """
-
-    pagination_class = Pagination
-    permission_classes = [
-        IsEtlUser,
-    ]
-    serializer_class = IngestibleCourseWithCourseRunsSerializer
-
-    def get_queryset(self):
-        """Get the queryset, with a bunch of prefetching for related data."""
-
-        queryset = Course.objects.select_related("page")
-        # Use Prefetch for reverse GenericRelation (products on CourseRun)
-        # 1. Get the ContentType object for the CourseRun model
-        courserun_content_type = ContentType.objects.get_for_model(CourseRun)
-        # 2. Create a Prefetch object to specify the queryset for the 'tags' relation
-        # This internal queryset only fetches products related to the CourseRun content type
-        courserun_product_queryset = Product.objects.filter(
-            content_type=courserun_content_type
-        )
-        # 3. Use prefetch_related on main queryset, referencing the reverse relation's name (e.g., 'products')
-        products_prefetch = Prefetch(
-            "products",
-            queryset=courserun_product_queryset,
-            to_attr="prefetched_products",
-        )
-        modes_prefetch = Prefetch(
-            "enrollment_modes",
-            to_attr="prefetched_enrollment_modes",
-        )
-        course_runs_prefetch = Prefetch(
-            "courseruns",
-            queryset=CourseRun.all_objects.order_by("id").prefetch_related(
-                modes_prefetch, products_prefetch
-            ),
-            to_attr="prefetched_courseruns",
-        )
-        queryset = queryset.prefetch_related(
-            "departments", "in_programs", course_runs_prefetch
-        )
-        queryset = queryset.annotate(
-            count_b2b_courseruns=Count("courseruns__b2b_contract__id")
-        )
-        queryset = queryset.annotate(count_courseruns=Count("courseruns"))
-        queryset = queryset.annotate(
-            verified_courserun_count=Count(
-                "courseruns__enrollment_modes",
-                filter=Q(
-                    courseruns__enrollment_modes__mode_slug=EDX_ENROLLMENT_VERIFIED_MODE
-                ),
-            )
-        )
-
-        return queryset.order_by("title").distinct()
