+++ openapi/specs/v2.yaml
@@ -3693,64 +3693,6 @@ paths:
               schema:
                 $ref: '#/components/schemas/CourseOutlineServerErrorResponse'
           description: ''
-  /api/v3/courses/ingestible/:
-    get:
-      operationId: courses_ingestible_list
-      description: |-
-        Largely a copy of the v2 CourseViewSet, but with changes for ETL processes.
-
-        The shape of the data returned should match the v2 CourseViewSet for the most
-        part, but we won't support lookups or filtering or anything and the prefetch
-        setup will be reduced (since this won't have to care about what the requesting
-        user's contracts are for instance).
-      parameters:
-      - name: page
-        required: false
-        in: query
-        description: A page number within the paginated result set.
-        schema:
-          type: integer
-      - name: page_size
-        required: false
-        in: query
-        description: Number of results to return per page.
-        schema:
-          type: integer
-      tags:
-      - courses
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/PaginatedIngestibleCourseWithCourseRunsSerializerV3List'
-          description: ''
-  /api/v3/courses/ingestible/{id}/:
-    get:
-      operationId: courses_ingestible_retrieve
-      description: |-
-        Largely a copy of the v2 CourseViewSet, but with changes for ETL processes.
-
-        The shape of the data returned should match the v2 CourseViewSet for the most
-        part, but we won't support lookups or filtering or anything and the prefetch
-        setup will be reduced (since this won't have to care about what the requesting
-        user's contracts are for instance).
-      parameters:
-      - in: path
-        name: id
-        schema:
-          type: integer
-        description: A unique integer value identifying this course.
-        required: true
-      tags:
-      - courses
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/IngestibleCourseWithCourseRunsSerializerV3'
-          description: ''
   /api/v3/courses/variant_runs/:
     get:
       operationId: course_variant_runs_v3
@@ -6601,152 +6543,6 @@ components:
       - key
       - text
       - title
-    IngestibleCourseWithCourseRunsSerializerV3:
-      type: object
-      description: Course model serializer - also serializes child course runs
-      properties:
-        id:
-          type: integer
-          readOnly: true
-        title:
-          type: string
-          maxLength: 255
-        readable_id:
-          type: string
-          pattern: ^[\w\-+:\.]+$
-          maxLength: 255
-        next_run_id:
-          type: integer
-          nullable: true
-          description: Get next run id
-          readOnly: true
-        departments:
-          type: array
-          items:
-            $ref: '#/components/schemas/Department'
-          readOnly: true
-        page:
-          allOf:
-          - $ref: '#/components/schemas/CoursePage'
-          readOnly: true
-          nullable: true
-        programs:
-          type: array
-          items:
-            $ref: '#/components/schemas/BaseProgram'
-          nullable: true
-          readOnly: true
-        topics:
-          type: array
-          items:
-            type: object
-            additionalProperties: {}
-          description: List topics of a course
-          readOnly: true
-        certificate_type:
-          type: string
-          description: Return the certificate type.
-          readOnly: true
-        certificate_available:
-          type: boolean
-          description: Return if there is a certificate available for the course.
-          readOnly: true
-        required_prerequisites:
-          type: boolean
-          description: |-
-            Check if the prerequisites field is populated in the course page CMS.
-            Returns:
-                bool: True when the prerequisites field is populated in the course page CMS.  False otherwise.
-          readOnly: true
-        duration:
-          type: string
-          description: Get the duration of the course from the course page CMS.
-          readOnly: true
-        min_weeks:
-          type: integer
-          nullable: true
-          description: Get the min weeks of the course from the CMS page.
-          readOnly: true
-        max_weeks:
-          type: integer
-          nullable: true
-          description: Get the max weeks of the course from the CMS page.
-          readOnly: true
-        min_price:
-          type: number
-          format: double
-          nullable: true
-          description: Get the min price of the product from the CMS page.
-          readOnly: true
-        max_price:
-          type: number
-          format: double
-          nullable: true
-          description: Get the max price of the product from the CMS page.
-          readOnly: true
-        time_commitment:
-          type: string
-          nullable: true
-          description: Get the time commitment of the course from the course page
-            CMS.
-          readOnly: true
-        availability:
-          type: string
-          description: Get course availability
-          readOnly: true
-        min_weekly_hours:
-          type: string
-          nullable: true
-          description: Get the min weekly hours of the course from the course page
-            CMS.
-          readOnly: true
-        max_weekly_hours:
-          type: string
-          nullable: true
-          description: Get the max weekly hours of the course from the course page
-            CMS.
-          readOnly: true
-        include_in_learn_catalog:
-          type: boolean
-          readOnly: true
-        ingest_content_files_for_ai:
-          type: boolean
-          readOnly: true
-        possible_variant_sets:
-          type: array
-          items:
-            $ref: '#/components/schemas/SupportedVariant'
-          readOnly: true
-        courseruns:
-          type: array
-          items:
-            $ref: '#/components/schemas/CourseRunV2'
-          readOnly: true
-      required:
-      - availability
-      - certificate_available
-      - certificate_type
-      - courseruns
-      - departments
-      - duration
-      - id
-      - include_in_learn_catalog
-      - ingest_content_files_for_ai
-      - max_price
-      - max_weekly_hours
-      - max_weeks
-      - min_price
-      - min_weekly_hours
-      - min_weeks
-      - next_run_id
-      - page
-      - possible_variant_sets
-      - programs
-      - readable_id
-      - required_prerequisites
-      - time_commitment
-      - title
-      - topics
     IntegrationTypeEnum:
       enum:
       - sso
@@ -7784,29 +7580,6 @@ components:
           type: array
           items:
             $ref: '#/components/schemas/FlexiblePriceTier'
-    PaginatedIngestibleCourseWithCourseRunsSerializerV3List:
-      type: object
-      required:
-      - count
-      - results
-      properties:
-        count:
-          type: integer
-          example: 123
-        next:
-          type: string
-          nullable: true
-          format: uri
-          example: http://api.example.org/accounts/?page=4
-        previous:
-          type: string
-          nullable: true
-          format: uri
-          example: http://api.example.org/accounts/?page=2
-        results:
-          type: array
-          items:
-            $ref: '#/components/schemas/IngestibleCourseWithCourseRunsSerializerV3'
     PaginatedOrderHistoryList:
       type: object
       required:
